+++ packages/pglite-socket/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @electric-sql/pglite-socket
 
+## 0.0.21
+
+### Patch Changes
+
+- 8a03647: Fix: Message buffering, connection handling, and concurrent connection support;
+
+## 0.0.20
+
+### Patch Changes
+
+- 54a4873: allow extensions to be loaded via '-e/--extensions <list>' cmd line parameter'
+- 45bff97: added pgcrypto extension
+- Updated dependencies [45bff97]
+- Updated dependencies [5ec474f]
+  - @electric-sql/pglite@0.3.15
+
 ## 0.0.19
 
 ### Patch Changes